fix(operator): test_svm_parser link + vmafx-operator bug audit
Bundles two independent correctness fixes.

(1) test_svm_parser Meson link break (pre-existing on master). The
target's source list omitted `../src/thread_locale.c`, so svm.cpp's
references to vmaf_thread_locale_push_c / vmaf_thread_locale_pop were
unresolved at link time. The sibling test_svm_api target already
includes thread_locale.c — same precedent applied here.

(2) cmd/vmafx-operator/ deep audit (Phase 4b kubebuilder operator,
ADR-0714). Three real defects:

- VmafxNode.probeHealthz deferred resp.Body.Close() without draining
  the body. Go net/http only returns connections to the keep-alive
  pool when the body is fully read; polling N VmafxNodes every 30s
  leaked one TCP connection per probe per node to the controller.
  Fixed via io.Copy(io.Discard, resp.Body) before Close.

- vmafx.dev/v1 CRD integer fields (VmafxJob.spec.priority,
  VmafxNode.spec.capacity, VmafxNode.status.assignedJobs,
  VmafxModelTraining.status.currentSamples,
  VmafxModelTraining.spec.checkpoint.minSamples) used Go `int`,
  violating Kubernetes API conventions — OpenAPI v3 has no
  architecture-dependent integer type. Widened to int32.

- Documented defaults (Backend=cpu, Priority=0, Capacity=1,
  Checkpoint.Interval=10m, Checkpoint.MinSamples=1000) were prose
  only — added +kubebuilder:default: markers on the Go types and
  default: keys in the CRD OpenAPI schemas so the apiserver actually
  applies them on admission. Helm CRD copies under
  deploy/helm/vmafx/crds/ resynced per cmd/vmafx-operator/AGENTS.md
  invariant #2.

Adds three standalone go-test regression tests
(TestProbeHealthzDrainsBody, TestProbeHealthzNon200StillDrains,
TestProbeHealthzTransportErrorReturnsFalse) that gate the body-drain
fix without requiring envtest binaries.

No ADR per CLAUDE §12 r8 (bug fixes).

+++ cmd/vmafx-operator/internal/controller/vmafxmodeltraining_controller_branch_test.go
@@ -69,7 +69,7 @@ var _ = Describe("VmafxModelTraining controller — branch coverage", func() {
 		var fetched vmafxv1.VmafxModelTraining
 		Expect(k8sClient.Get(ctx, nn, &fetched)).To(Succeed())
 		fetched.Status.Phase = vmafxv1.VmafxModelTrainingPhaseRunning
-		fetched.Status.CurrentSamples = 42
+		fetched.Status.CurrentSamples = int32(42)
 		fetched.Status.ModelVersion = "sha256:deadbeef"
 		Expect(k8sClient.Status().Update(ctx, &fetched)).To(Succeed())
 
@@ -86,7 +86,7 @@ var _ = Describe("VmafxModelTraining controller — branch coverage", func() {
 		var after vmafxv1.VmafxModelTraining
 		Expect(k8sClient.Get(ctx, nn, &after)).To(Succeed())
 		Expect(after.Status.Phase).To(Equal(vmafxv1.VmafxModelTrainingPhaseRunning))
-		Expect(after.Status.CurrentSamples).To(Equal(42))
+		Expect(after.Status.CurrentSamples).To(Equal(int32(42)))
 		Expect(after.Status.ModelVersion).To(Equal("sha256:deadbeef"))
 	})
 })
